refactor(db): extract migrations into store_migrations.go — part 1 of the store.go split (#1037)
Pure same-package relocation (package db): moves `var migrations` (92 entries,
1,549 lines), `Migrate`, and `applyMigration` out of the 9,378-line store.go into
a new store_migrations.go. First and cheapest step of the #1029-audited store.go
split; the domain-file splits (store_jobs/skillopt/templates/tasks/types/agents/
locks) follow in subsequent parts.

The migrations slice is POSITIONALLY versioned (array index = version number) —
moved verbatim, never renumbered/reordered/squashed.

Byte-identical / no behavior change, verified deterministically:
- migrations slice byte-identical (1,549 lines, content + order); 92 entries → 92.
- Migrate (270B) + applyMigration (887B) bodies byte-identical to HEAD.
- store.go: 0 added lines (pure removal); top-level declaration set unchanged
  (400 → 400, union across both files identical).
- Gate: go build/vet/generate clean, gofmt clean, go test ./internal/db/
  ./internal/cli/ ./internal/workflow/ ./internal/daemon/ green (incl. schema
  migration + idempotency coverage).
